requirements.txt
flask
python-dotenv
groq
pytesseract
Pillow
pdf2image
numpy
